requirements.txt
openai
anthropic
python-dotenv
tiktoken
numpy
matplotlib
networkx
